[fix] 버그 수정, dao 조건추가
nickname 중복확인시 deletedAt is null 조건 추가
내가 등록한, 저장한 게시글 불러올시 deletedAt is null 조건 추가
cafe 검색시 해당 카페의 가장 최신 포스트의 postId를 같이 응답
지도 리스트업시 정수로만 들어가는 버그 수정

+++ src/controllers/cafesController.js
@@ -25,10 +25,10 @@ exports.searchCafe = async (req, res) => {
 
 exports.cafesInMap = async (req, res) => {
   const currentMapRange = {
-    swLng: parseInt(req.query.swLng),
-    swLat: parseInt(req.query.swLat),
-    neLng: parseInt(req.query.neLng),
-    neLat: parseInt(req.query.neLat),
+    swLng: parseFloat(req.query.swLng),
+    swLat: parseFloat(req.query.swLat),
+    neLng: parseFloat(req.query.neLng),
+    neLat: parseFloat(req.query.neLat),
   };
 
   const currentLocation = {
